Refatoring, and added human and robot class
